prototypes/Video-To-Frames/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Video-To-Frames/all_videos_to_images.py: 7
# Video-To-Frames/video_to_frame_categories.py: 6
ffmpeg_python == 0.2.0
